backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
httpx>=0.26.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
cryptography>=42.0.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
pandas>=2.2.0
openpyxl>=3.1.0
pypdf>=4.0.0
greenlet>=3.0.0
PyJWT>=2.8.0
bcrypt>=4.0.0
email-validator>=2.0.0
